refactor: corsConfig.ts 를 lib/config 폴더로 이동
관련 router 파일들도 모두 수정
