+++ server/src/Stories/Stories.router.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import cors from 'cors';
-import options from '../../lib/corsConfig';
+import options from '../../lib/config/corsConfig';
 import { getAllStoriesByProject } from './Stories.controller';
 
 const router = express.Router();
